Enrollment Service/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.28
psycopg[binary]==3.2.13
python-dotenv==1.0.1
pydantic==2.6.4
pydantic-settings==2.2.1
PyJWT==2.8.0
passlib==1.7.4
python-multipart==0.0.9
httpx==0.27.2
